+++ psgi/test/metadata.rb
@@ -6,3 +6,4 @@ description      'Installs/Configures test'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 depends          'cpan'
+depends          'psgi'