+++ psgi/CHANGELOG.md
@@ -4,7 +4,7 @@ This file is used to list changes made in each version of psgi.
 
 ### 0.0.7
 * berkshelf / minitest chef infrastructure has been added
-* centos now is supported - todo
+* default value for `proc_manager` is FCGI::ProcManager
 
 ## 0.0.6
 * delete old files